+++ alf/networks/__init__.py
@@ -18,7 +18,7 @@ from .critic_networks import *
 from .encoding_networks import *
 from .network import Network, NaiveParallelNetwork
 from .ou_process import OUProcess
-from .preprocessors import PreprocessorNetwork
+from .preprocessor_networks import PreprocessorNetwork
 from .projection_networks import *
 from .q_networks import *
 from .value_networks import *
