+++ alf/networks/actor_distribution_networks.py
@@ -21,7 +21,7 @@ import torch.nn as nn
 import alf.nest as nest
 from .encoding_networks import EncodingNetwork, LSTMEncodingNetwork
 from .projection_networks import NormalProjectionNetwork, CategoricalProjectionNetwork
-from .preprocessors import PreprocessorNetwork
+from .preprocessor_networks import PreprocessorNetwork
 from alf.tensor_specs import BoundedTensorSpec, TensorSpec
 from alf.networks.network import Network
