+++ alf/networks/actor_networks.py
@@ -21,7 +21,7 @@ import torch
 import torch.nn as nn
 
 from .encoding_networks import EncodingNetwork, LSTMEncodingNetwork
-from .preprocessors import PreprocessorNetwork
+from .preprocessor_networks import PreprocessorNetwork
 import alf.layers as layers
 import alf.nest as nest
 from alf.initializers import variance_scaling_init
