+++ alf/networks/preprocessors.py
@@ -129,109 +129,3 @@ class EmbeddingPreprocessor(InputPreprocessor):
         ret = self._embedding_net(tensor)
         # EncodingNetwork returns a pair
         return (ret if self._input_tensor_spec.is_discrete else ret[0])
-
-
-class PreprocessorNetwork(Network):
-    """A base class for networks with input processing need."""
-
-    def __init__(self,
-                 input_tensor_spec,
-                 input_preprocessors=None,
-                 preprocessing_combiner=None,
-                 name="PreprocessorNetwork"):
-        """
-        Args:
-            input_tensor_spec (nested TensorSpec): the (nested) tensor spec of
-                the input. If nested, then ``preprocessing_combiner`` must not be
-                None.
-            input_preprocessors (nested InputPreprocessor): a nest of
-                ``InputPreprocessor``, each of which will be applied to the
-                corresponding input. If not None, then it must have the same
-                structure with ``input_tensor_spec``. If any element is None, then
-                it will be treated as math_ops.identity. This arg is helpful if
-                you want to have separate preprocessings for different inputs by
-                configuring a gin file without changing the code. For example,
-                embedding a discrete input before concatenating it to another
-                continuous vector.
-            preprocessing_combiner (NestCombiner): preprocessing called on
-                complex inputs. Note that this combiner must also accept
-                ``input_tensor_spec`` as the input to compute the processed
-                tensor spec. For example, see `alf.nest.utils.NestConcat`. This
-                arg is helpful if you want to combine inputs by configuring a
-                gin file without changing the code.
-            name (str): name of the network
-        """
-        super().__init__(input_tensor_spec, name)
-
-        # make sure the network holds the parameters of any trainable input
-        # preprocessor
-        self._input_preprocessor_modules = nn.ModuleList()
-
-        def _get_preprocessed_spec(preproc, spec):
-            if not isinstance(preproc, InputPreprocessor):
-                # In this case we just assume the spec won't change after the
-                # preprocessing. If it does change, then you should consider
-                # defining an `InputPreprocessor` instead.
-                return spec
-            return preproc(spec)
-
-        self._input_preprocessors = None
-        if input_preprocessors is not None:
-            input_tensor_spec = alf.nest.map_structure(
-                _get_preprocessed_spec, input_preprocessors, input_tensor_spec)
-
-            def _return_or_copy_preprocessor(preproc):
-                if preproc is None:
-                    # allow None as a placeholder in the nest
-                    preproc = math_ops.identity
-                elif isinstance(preproc, InputPreprocessor):
-                    preproc = preproc.copy()
-                    self._input_preprocessor_modules.append(preproc)
-                return preproc
-
-            self._input_preprocessors = alf.nest.map_structure(
-                _return_or_copy_preprocessor, input_preprocessors)
-
-        self._preprocessing_combiner = preprocessing_combiner
-        if alf.nest.is_nested(input_tensor_spec):
-            assert preprocessing_combiner is not None, \
-                ("When a nested input tensor spec is provided, a input " +
-                "preprocessing combiner must also be provided!")
-            input_tensor_spec = preprocessing_combiner(input_tensor_spec)
-        else:
-            assert isinstance(input_tensor_spec, TensorSpec), \
-                "The spec must be an instance of TensorSpec!"
-            self._preprocessing_combiner = math_ops.identity
-
-        # This input spec is the final resulting spec after input preprocessors
-        # and the nest combiner.
-        self._processed_input_tensor_spec = input_tensor_spec
-
-    def forward(self, inputs, state=(), min_outer_rank=1, max_outer_rank=1):
-        """Preprocessing nested inputs.
-
-        Args:
-            inputs (nested Tensor): inputs to the network
-            state (nested Tensor): RNN state of the network
-            min_outer_rank (int): the minimal outer rank allowed
-            max_outer_rank (int): the maximal outer rank allowed
-        Returns:
-            Tensor: tensor after preprocessing.
-        """
-        if self._input_preprocessors:
-            inputs = alf.nest.map_structure(
-                lambda preproc, tensor: preproc(tensor),
-                self._input_preprocessors, inputs)
-        proc_inputs = self._preprocessing_combiner(inputs)
-        outer_rank = get_outer_rank(proc_inputs,
-                                    self._processed_input_tensor_spec)
-        assert min_outer_rank <= outer_rank <= max_outer_rank, \
-            ("Only supports {}<=outer_rank<={}! ".format(min_outer_rank, max_outer_rank)
-            + "After preprocessing: inputs size {} vs. input tensor spec {}".format(
-                proc_inputs.size(), self._processed_input_tensor_spec)
-            + "\n Make sure that you have provided the right input preprocessors"
-            + " and nest combiner!\n"
-            + "Before preprocessing: inputs size {} vs. input tensor spec {}".format(
-                alf.nest.map_structure(lambda tensor: tensor.size(), inputs),
-                self._input_tensor_spec))
-        return proc_inputs, state
